hey-translate/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

customtkinter==5.2.2
deepl==1.22.0
pillow==11.2.1
platformdirs==4.3.8
realtimestt==0.3.104
sounddevice==0.5.2
torch==2.7.0+cu128
torchaudio==2.7.0
